Fix for issue #8
See details at https://github.com/suprb/Grid-A-Licious/issues/8.
When the container element was hidden the width would return as 0.
And becuase width was set to 0 no collumns were add so the elements
were not being added.  This work around replaces the 'display' poperty to
'block', then finds the width, and then replaces the property back to where
it was set.